requirements.txt
pdf2image
pytesseract
opencv-python
numpy
pandas
Pillow
fastapi[all]
unicorn